whatsapp_bot/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
audioread==3.0.1
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
fastapi==0.110.3
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
lazy_loader==0.4
llvmlite==0.42.0
MarkupSafe==2.1.5
msgpack==1.0.8
multidict==6.0.5
numba==0.59.1
numpy==1.26.4
opencv-python==4.9.0.80
opuslib==3.0.1
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
pooch==1.8.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyJWT==2.8.0
pyngrok==7.1.6
PyOgg==0.6.14a1
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.29
starlette==0.37.2
threadpoolctl==3.5.0
tqdm==4.66.2
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
yarl==1.9.4
